DDF-2196: Separate admin duties via Admin Console (#860)
- Created policy extension to authorize users to features and services
	- Made AbstractIntegrationTest use the system subject to make calls to the ServiceManager via proxy

+++ platform/admin/core/admin-core-appservice/src/main/java/org/codice/ddf/admin/application/service/impl/ApplicationServiceImpl.java
@@ -38,6 +38,7 @@ import org.apache.karaf.features.Feature;
 import org.apache.karaf.features.FeaturesService;
 import org.apache.karaf.features.FeaturesService.Option;
 import org.apache.karaf.features.Repository;
+import org.apache.shiro.SecurityUtils;
 import org.codice.ddf.admin.application.rest.model.FeatureDetails;
 import org.codice.ddf.admin.application.service.Application;
 import org.codice.ddf.admin.application.service.ApplicationNode;
@@ -57,6 +58,11 @@ import org.osgi.service.cm.Configuration;
 import org.osgi.service.cm.ConfigurationAdmin;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.google.common.collect.Sets;
+
+import ddf.security.permission.KeyValueCollectionPermission;
+import ddf.security.permission.KeyValuePermission;
 
 /**
  * Implementation of the ApplicationService. Uses the karaf features service and
@@ -137,7 +143,7 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
             Application newApp = new ApplicationImpl(repos[i]);
             try {
                 if (!ignoredApplicationNames.contains(newApp.getName()) && newApp.getFeatures()
-                        .size() > 0) {
+                        .size() > 0 && isPermittedToViewFeature(newApp.getName())) {
                     applications.add(newApp);
                 }
             } catch (ApplicationServiceException ase) {
@@ -796,7 +802,7 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
     @Override
     public void removeApplication(Application application) throws ApplicationServiceException {
         try {
-            if (application != null) {
+            if (application != null && isPermittedToViewFeature(application.getName())) {
                 uninstallAllFeatures(application);
                 featuresService.removeRepository(application.getURI(), false);
             }
@@ -844,7 +850,7 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
         try {
             Set<Feature> features = application.getFeatures();
             for (Feature feature : features) {
-                if (featuresService.isInstalled(feature)) {
+                if (featuresService.isInstalled(feature) && isPermittedToViewFeature(feature.getName())) {
                     try {
                         featuresService.uninstallFeature(feature.getName(),
                                 feature.getVersion(),
@@ -868,7 +874,9 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
         List<FeatureDetails> features = new ArrayList<FeatureDetails>();
         try {
             for (Feature feature : featuresService.listFeatures()) {
-                features.add(getFeatureView(feature));
+                if (isPermittedToViewFeature(feature.getName())) {
+                    features.add(getFeatureView(feature));
+                }
             }
         } catch (Exception ex) {
             LOGGER.warn("Could not obtain all features.", ex);
@@ -900,6 +908,7 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
     public List<FeatureDetails> findApplicationFeatures(String applicationName) {
         List<FeatureDetails> features = getRepositoryFeatures(applicationName).stream()
                 .filter(feature -> !isAppInFeatureList(feature, applicationName))
+                .filter(feature -> isPermittedToViewFeature(feature.getName()))
                 .map(this::getFeatureView)
                 .collect(Collectors.toList());
         return features;
@@ -955,6 +964,12 @@ public class ApplicationServiceImpl implements ApplicationService, ServiceListen
                 context.removeServiceListener(this);
             }
         }
+    }
+
+    public boolean isPermittedToViewFeature(String featureName){
+        KeyValueCollectionPermission serviceToCheck = new KeyValueCollectionPermission("view-feature.name",
+                new KeyValuePermission("feature.name", Sets.newHashSet(featureName)));
+        return SecurityUtils.getSubject().isPermitted(serviceToCheck);
     }
 
     /**
